backend/requirements.txt
# Core dependencies
python-dotenv
django
djangorestframework
azure-search-documents 
azure-ai-textanalytics 
python-dotenv
djangorestframework-simplejwt
django-cors-headers
drf-spectacular

# Testing dependencies
pytest
pytest-django
pytest-cov
coverage

